create the validate package and move validate_enc into
